major refactor + enhancements
