+++ packages/ui/app/_components/dashboards/InfoRows.tsx
@@ -1,7 +1,8 @@
 /* eslint-disable @next/next/no-img-element */
 'use client';
 import dynamic from 'next/dynamic';
-import type { Dispatch, SetStateAction } from 'react';
+import { useMemo, type Dispatch, type SetStateAction } from 'react';
+import type { FlywheelReward } from 'types/dist';
 import type { Address } from 'viem';
 
 import { getAssetName } from '../../util/utils';
@@ -9,8 +10,11 @@ import { getAssetName } from '../../util/utils';
 const Rewards = dynamic(() => import('../markets/Rewards'), {
   ssr: false
 });
+import BorrowPopover from '../markets/BorrowPopover';
+import SupplyPopover from '../markets/SupplyPopover';
 import { PopupMode } from '../popup/page';
 
+import { FLYWHEEL_TYPE_MAP, pools } from '@ui/constants/index';
 import { multipliers } from '@ui/utils/multipliers';
 
 export enum InfoMode {
@@ -23,12 +27,13 @@ export type InfoRowsProps = {
   apr: string;
   asset: string;
   collateralApr: string;
-  cToken: string;
+  cToken: `0x${string}`;
   logo: string;
   pool: string;
   membership: boolean;
   mode: InfoMode;
   comptrollerAddress: Address;
+  rewards: FlywheelReward[];
   selectedChain: number;
   setPopupMode: Dispatch<SetStateAction<PopupMode | undefined>>;
   setSelectedSymbol: Dispatch<SetStateAction<string>>;
@@ -47,8 +52,33 @@ const InfoRows = ({
   selectedChain,
   cToken,
   comptrollerAddress,
-  pool
+  pool,
+  rewards
 }: InfoRowsProps) => {
+  const supplyRewards = useMemo(
+    () =>
+      rewards?.filter((reward) =>
+        FLYWHEEL_TYPE_MAP[+selectedChain]?.supply?.includes(
+          (reward as FlywheelReward).flywheel
+        )
+      ),
+    [selectedChain, rewards]
+  );
+
+  const borrowRewards = useMemo(
+    () =>
+      rewards?.filter((reward) =>
+        FLYWHEEL_TYPE_MAP[+selectedChain]?.borrow?.includes(
+          (reward as FlywheelReward).flywheel
+        )
+      ),
+    [selectedChain, rewards]
+  );
+  const totalBorrowRewardsAPR = useMemo(
+    () =>
+      borrowRewards?.reduce((acc, reward) => acc + (reward.apy ?? 0), 0) ?? 0,
+    [borrowRewards]
+  );
   return (
     <div
       className={`w-full hover:bg-graylite transition-all duration-200 ease-linear bg-grayUnselect rounded-xl mb-3 px-2  gap-x-1 md:grid  grid-cols-5  py-4 text-xs text-white/80 font-semibold text-center items-center relative ${
@@ -87,7 +117,37 @@ const InfoRows = ({
         <span className="text-white/40 font-semibold mr-2 md:hidden text-left">
           {mode === InfoMode.SUPPLY ? 'SUPPLY' : 'BORROW'} APR:
         </span>
-        {apr}
+        <div
+          className={` mb-2 popover-container relative flex md:flex-col items-center justify-between md:justify-center cursor-pointer`}
+        >
+          {apr}
+          {mode === InfoMode.SUPPLY ? (
+            <>
+              <SupplyPopover
+                asset={asset}
+                supplyAPR={parseFloat(apr.replace('%', ''))}
+                rewards={supplyRewards}
+                dropdownSelectedChain={+selectedChain}
+                selectedPoolId={pool}
+                cToken={cToken}
+                pool={comptrollerAddress}
+              />
+            </>
+          ) : (
+            <>
+              <BorrowPopover
+                asset={asset}
+                borrowAPR={parseFloat(apr.replace('%', ''))}
+                rewardsAPR={totalBorrowRewardsAPR}
+                dropdownSelectedChain={+selectedChain}
+                selectedPoolId={pool}
+                cToken={cToken}
+                pool={comptrollerAddress}
+                rewards={borrowRewards}
+              />
+            </>
+          )}
+        </div>
       </h3>
       {multipliers[selectedChain]?.[pool]?.[asset]?.borrow?.flywheel &&
       mode == InfoMode.BORROW ? (
@@ -126,7 +186,7 @@ const InfoRows = ({
         </button>
 
         <button
-          className={`w-full uppercase bg-lime rounded-lg text-black py-1.5 px-3`}
+          className={`w-full uppercase ${pools[+selectedChain].text} ${pools[+selectedChain].bg} rounded-lg text-black py-1.5 px-3`}
           onClick={() => {
             setSelectedSymbol(asset);
             setPopupMode(
