+++ packages/ui/app/dashboard/page.tsx
@@ -3,11 +3,12 @@
 
 import millify from 'millify';
 import Link from 'next/link';
-import { usePathname, useSearchParams, useRouter } from 'next/navigation';
+import { usePathname, useSearchParams } from 'next/navigation';
 import { useEffect, useMemo, useRef, useState } from 'react';
+import type { FlywheelReward } from 'types/dist';
 import { type Address, formatEther, formatUnits, parseEther } from 'viem';
-import { base } from 'viem/chains';
-import { useChainId } from 'wagmi';
+// import { base } from 'viem/chains';
+// import { useChainId } from 'wagmi';
 
 import InfoRows, { InfoMode } from '../_components/dashboards/InfoRows';
 import NetworkSelector from '../_components/markets/NetworkSelector';
@@ -16,9 +17,9 @@ import type { PopupMode } from '../_components/popup/page';
 import Popup from '../_components/popup/page';
 import ResultHandler from '../_components/ResultHandler';
 
-import { pools, REWARDS_TO_SYMBOL } from '@ui/constants/index';
+import { pools } from '@ui/constants/index';
 import { useMultiIonic } from '@ui/context/MultiIonicContext';
-import { useSdk } from '@ui/hooks/fuse/useSdk';
+// import { useSdk } from '@ui/hooks/fuse/useSdk';
 import { useCurrentLeverageRatios } from '@ui/hooks/leverage/useCurrentLeverageRatio';
 import { usePositionsInfo } from '@ui/hooks/leverage/usePositionInfo';
 import { usePositionsQuery } from '@ui/hooks/leverage/usePositions';
@@ -36,10 +37,10 @@ import {
   usePointsForSupplyBaseMain,
   usePointsForSupplyModeMain
 } from '@ui/hooks/usePointsQueries';
+import { useRewards } from '@ui/hooks/useRewards';
 import { useTotalSupplyAPYs } from '@ui/hooks/useTotalSupplyAPYs';
 import { useUserNetApr } from '@ui/hooks/useUserNetApr';
 import type { MarketData } from '@ui/types/TokensDataMap';
-import { handleSwitchOriginChain } from '@ui/utils/NetworkChecker';
 import { getBlockTimePerMinuteByChainId } from '@ui/utils/networkData';
 
 export default function Dashboard() {
@@ -219,6 +220,8 @@ export default function Dashboard() {
   //     marketData?.comptroller ?? '',
   //     +chain
   //   );
+
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
   const totalPoints = useMemo<number>(() => {
     if (
       supplyPointsNative &&
@@ -358,6 +361,21 @@ export default function Dashboard() {
   //   }
   //   return marketData?.assets.map(() => '0.00%') ?? [];
   // }, [borrowCaps, marketData]);
+
+  const { data: rewards } = useRewards({
+    chainId: +chain,
+    poolId: pool
+  });
+
+  const allChains: number[] = Object.keys(pools).map(Number);
+  const { data: claimableRewardsAcrossAllChains } =
+    useAllClaimableRewards(allChains);
+  const totalRewardsAcrossAllChains =
+    claimableRewardsAcrossAllChains?.reduce(
+      (acc, reward) => acc + reward.amount,
+      0n
+    ) ?? 0n;
+  // console.log(claimableRewardsAcrossAllChains , totalRewardsAcrossAllChains)
 
   return (
     <>
@@ -506,10 +524,13 @@ export default function Dashboard() {
                 }
                 width="24"
               >
-                <span>
-                  {Math.round(totalPoints).toLocaleString('en-us', {
+                <span className={`flex items-center justify-center gap-1`}>
+                  {Math.round(
+                    +formatEther(totalRewardsAcrossAllChains)
+                  ).toLocaleString('en-us', {
                     maximumFractionDigits: 0
-                  })}
+                  })}{' '}
+                  <span className={`text-[8px] text-white/50`}>(ION+RSR)</span>
                 </span>
               </ResultHandler>
             </div>
@@ -517,7 +538,7 @@ export default function Dashboard() {
               className={`w-full rounded-md bg-accent text-black py-2 px-6 text-center text-xs mt-auto  `}
               href={`/points`}
             >
-              VIEW POINTS
+              CLAIM ALL REWARDS
             </Link>
           </div>
         </div>
@@ -666,6 +687,15 @@ export default function Dashboard() {
                         marketData?.comptroller ?? ('' as Address)
                       }
                       pool={pool}
+                      rewards={
+                        (rewards?.[asset?.cToken]?.map((r) => ({
+                          ...r,
+                          apy:
+                            typeof r.apy !== 'undefined'
+                              ? r.apy * 100
+                              : undefined
+                        })) as FlywheelReward[]) ?? []
+                      }
                       selectedChain={+chain}
                       setPopupMode={setPopupMode}
                       setSelectedSymbol={setSelectedSymbol}
@@ -743,6 +773,15 @@ export default function Dashboard() {
                         marketData?.comptroller ?? ('' as Address)
                       }
                       pool={pool}
+                      rewards={
+                        (rewards?.[asset?.cToken]?.map((r) => ({
+                          ...r,
+                          apy:
+                            typeof r.apy !== 'undefined'
+                              ? r.apy * 100
+                              : undefined
+                        })) as FlywheelReward[]) ?? []
+                      }
                       key={`supply-row-${asset.underlyingSymbol}`}
                       logo={`/img/symbols/32/color/${asset.underlyingSymbol.toLowerCase()}.png`}
                       membership={asset.membership}
@@ -956,92 +995,3 @@ export default function Dashboard() {
     </>
   );
 }
-
-interface IClaimAllBaseRewards {
-  chain: number;
-}
-const ClaimAllBaseRewards = ({ chain }: IClaimAllBaseRewards) => {
-  const { data: rewards } = useAllClaimableRewards([+chain]);
-  const [loading, setLoading] = useState<boolean>(false);
-  const sdk = useSdk(+chain);
-  const chainId = useChainId();
-  async function claimAll() {
-    try {
-      const result = await handleSwitchOriginChain(+chain, chainId);
-      if (!result) return;
-      setLoading(true);
-      await sdk?.claimAllRewards();
-      setLoading(false);
-    } catch (err) {
-      console.warn(err);
-    } finally {
-      setLoading(false);
-    }
-  }
-  // console.log(rewards);
-  const totalRewards =
-    rewards?.reduce((acc, reward) => acc + reward.amount, 0n) ?? 0n;
-
-  const getSymbol = (chain: number, token: string) =>
-    //@ts-ignore
-    REWARDS_TO_SYMBOL[chain][token];
-
-  const router = useRouter();
-
-  const hrefTOStake = () => router.push('/stake');
-  return (
-    <div
-      className={`  bg-grayone px-3 py-3 rounded-lg flex flex-col items-start justify-start`}
-    >
-      <div className={` mb-2 w-full grid grid-cols-3 items-center`}>
-        <p className="text-white/60 text-md">Emissions </p>
-        <div className="flex items-center justify-start gap-1 col-start-3">
-          {rewards && totalRewards > 0n ? (
-            rewards?.map((reward, idx) => (
-              <img
-                alt="icon"
-                key={idx}
-                className="size-6 rounded mr-1  inline-block"
-                src={`/img/symbols/32/color/${getSymbol(+chain, reward.rewardToken)}.png`}
-              />
-            ))
-          ) : (
-            <p className="text-white/60 text-xs col-start-3">No Emissions </p>
-          )}
-        </div>
-        {/* {rewards ? "" : } */}
-      </div>
-      <div className=" w-full grid grid-cols-3 ">
-        <div className="flex items-center justify-start gap-4 ">
-          {rewards?.map((reward, idx) => (
-            <div
-              key={idx}
-              className={`text-white/80 text-sm flex gap-2 items-center justify-start `}
-            >
-              <span>{getSymbol(+chain, reward.rewardToken)}</span>
-              <span className={`text-accent text-sm`}>
-                {Number(formatEther(reward.amount)).toLocaleString('en-US', {
-                  maximumFractionDigits: 1
-                })}
-              </span>
-            </div>
-          ))}
-        </div>
-        <button
-          className={`rounded-md bg-accent text-black disabled:bg-accent/50 py-0.5 px-3 uppercase truncate text-[11px]  font-semibold col-start-3 `}
-          onClick={totalRewards > 0n ? claimAll : hrefTOStake}
-          disabled={loading && totalRewards > 0n}
-        >
-          <ResultHandler
-            isLoading={loading}
-            height="20"
-            width="20"
-            color={'#000000'}
-          >
-            {totalRewards > 0n ? 'Claim All Rewards' : 'Stake Rewards'}
-          </ResultHandler>
-        </button>
-      </div>
-    </div>
-  );
-};
